requirements.txt
numpy == 1.18.5
matplotlib == 3.3.2
dlib == 19.21.0
tensorflow == 2.3.1
Keras == 2.4.3
Keras-Applications == 1.0.8
Keras-Preprocessing  ==  1.1.2
keras-vggface == 0.6
Pillow == 7.2.0
opencv-python == 4.4.0.44
mtcnn == 0.1.0
